+++ scripts/publishPackages.js
@@ -5,7 +5,6 @@ const semver = require('semver');
 const memoize = require('lodash/memoize');
 const get = require('lodash/get');
 const lernaPackages = require('lerna-packages');
-const deployExamples = require('./deployExamples');
 
 const LATEST_TAG = 'latest';
 const NEXT_TAG = 'next';
@@ -100,55 +99,21 @@ function publishPackages() {
     .forEach(p => release(p));
 }
 
-function publishExamples() {
-  if (publishedPackages.length) {
-    const packagesWithExamples = [
-      {
-        packageName: 'wix-rich-content-editor',
-        exampleName: 'rich-content-editor',
-        examplePath: 'examples/editor',
-      },
-      {
-        packageName: 'wix-rich-content-viewer',
-        exampleName: 'rich-content-viewer',
-        examplePath: 'examples/viewer',
-      },
-    ];
-
-    const examplesToDeploy = [];
-    packagesWithExamples.forEach(packageWithExample => {
-      const pkg = publishedPackages.find(p => packageWithExample.packageName === p.name);
-      if (pkg) {
-        const { exampleName: name, examplePath: path } = packageWithExample;
-        const { version } = pkg;
-        examplesToDeploy.push({
-          name,
-          path,
-          version,
-        });
-      }
-    });
-    deployExamples(examplesToDeploy);
-  } else {
-    console.log('No examples to publish');
-  }
-}
-
 function run() {
-  if (process.env.TRAVIS_BRANCH !== 'master') {
-    let skipReason;
-    if (process.env.CI) {
-      skipReason = 'Not on master branch';
-    } else {
-      skipReason = 'Not in CI';
-    }
-    console.log(chalk.yellow(`${skipReason} - skipping publish`));
+  let skip;
+  const { TRAVIS_BRANCH, CI } = process.env;
+  if (TRAVIS_BRANCH !== 'master') {
+    skip = 'Not on master branch';
+  } else if (!CI) {
+    skip = 'Not in CI';
+  }
+  if (skip) {
+    console.log(chalk.yellow(`${skip} - skipping publish`));
     return false;
   }
 
   createNpmRc();
   publishPackages();
-  // publishExamples();
 }
 
 run();